# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Bmstu Open IT Platform",
        "title": "Forecast system",
        "contact": {
            "name": "API Support",
            "url": "https://vk.com/b.larkin",
            "email": "borislarkin18@mail.ru"
        },
        "license": {
            "name": "AS IS (NO WARRANTY)"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/forecast/add": {
            "post": {
                "description": "very very friendly response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecasts"
                ],
                "summary": "Add forecast to the list",
                "parameters": [
                    {
                        "description": "New forecast data",
                        "name": "forecast",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ForecastRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.ForecastRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ds.Forecasts"
                        }
                    }
                }
            }
        },
        "/forecast/delete/{id}": {
            "delete": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecasts"
                ],
                "summary": "Delete a specified forecast by its ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/forecast/edit/{id}": {
            "put": {
                "description": "very very friendly response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecasts"
                ],
                "summary": "Edit forecast",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New forecast data",
                        "name": "forecast",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ForecastRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.ForecastRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ds.Forecasts"
                        }
                    }
                }
            }
        },
...
        "/forecast/{id}": {
            "get": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecasts"
                ],
                "summary": "Get a specified forecast by its ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.ForecastResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/forecast/{id}/add_picture": {
            "post": {
                "description": "very very friendly response",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Forecasts"
                ],
                "summary": "Add image to specified forecast",
                "parameters": [
                    {
                        "type": "file",
                        "description": "New image for the forecast",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Forecast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth Bearer token header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/forecasts": {
            "get": {
                "description": "very very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forecasts"
                ],
                "summary": "Show all available forecasts filtered by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name filter",
                        "name": "forecast_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.GetForecastsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.ForecastRequest": {
            "type": "object",
            "required": [
                "color",
                "descr",
                "ext_desc",
                "measure_type",
                "short_title",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "ext_desc": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "measure_type": {
                    "type": "string"
                },
                "short_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.ForecastResponse": {
            "type": "object",
            "required": [
                "color",
                "descr",
                "ext_desc",
                "forecast_id",
                "image",
                "measure_type",
                "short_title",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "ext_desc": {
                    "type": "string"
                },
                "forecast_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "measure_type": {
                    "type": "string"
                },
                "short_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "ds.Forecasts": {
            "type": "object",
            "required": [
                "color",
                "descr",
                "ext_desc",
                "forecast_id",
                "measure_type",
                "short_title",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string"
                },
                "descr": {
                    "type": "string"
                },
                "ext_desc": {
                    "type": "string"
                },
                "forecast_id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "measure_type": {
                    "type": "string"
                },
                "short_title": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.GetForecastsResponse": {
            "type": "object",
            "required": [
                "forecasts",
                "prediction_id",
                "prediction_size"
            ],
            "properties": {
                "forecasts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Forecasts"
                    }
                },
                "forecasts_empty": {
                    "type": "boolean"
                },
                "prediction_id": {
                    "type": "string"
                },
                "prediction_size": {
                    "type": "integer"
                }
            }
        },
...
    }
}